lapbot/src/qa_data/transformation/requirement.txt
python-dotenv
google-generativeai
pandas
tqdm